roles and data ownership

+++ src/tasks/tasks.controller.ts
@@ -10,6 +10,7 @@ import {
   HttpStatus,
   HttpCode,
   BadRequestException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { TasksService } from './tasks.service';
 import { CreateTaskDto } from './create-task.dto';
@@ -22,6 +23,7 @@ import { FindTaskParams } from './find-task.params';
 import { Query } from '@nestjs/common';
 import { PaginationParams } from 'src/common/pagination.params';
 import { PaginationResponse } from 'src/common/pagination.response';
+import { CurrentUserId } from 'src/users/decorators/current-user-id.decorator';
 
 @Controller('tasks')
 export class TasksController {
@@ -31,8 +33,13 @@ export class TasksController {
   public async findAll(
     @Query() filters: FindTaskParams,
     @Query() pagination: PaginationParams,
+    @CurrentUserId() userId: string,
   ): Promise<PaginationResponse<Task>> {
-    const [tasks, count] = await this.tasksService.findAll(filters, pagination);
+    const [tasks, count] = await this.tasksService.findAll(
+      filters,
+      pagination,
+      userId,
+    );
     return {
       data: tasks,
       meta: {
@@ -44,13 +51,21 @@ export class TasksController {
   }
 
   @Get('/:id')
-  public async findOne(@Param() params: FindOneParams): Promise<Task> {
-    return await this.findOneOrFail(params.id);
+  public async findOne(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+    return task;
   }
 
   @Post()
-  public async create(@Body() createTaskDto: CreateTaskDto): Promise<Task> {
-    return await this.tasksService.createTask(createTaskDto);
+  public async create(
+    @Body() createTaskDto: CreateTaskDto,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    return await this.tasksService.createTask({ ...createTaskDto, userId });
   }
 
   /*  @Patch('/:id/status')
@@ -67,8 +82,11 @@ export class TasksController {
   public async updateTask(
     @Param() params: FindOneParams,
     @Body() updateTaskDto: UpdateTaskDto,
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+
     try {
       return await this.tasksService.updateTask(task, updateTaskDto);
     } catch (error) {
@@ -81,8 +99,12 @@ export class TasksController {
 
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  public async deleteTask(@Param() params: FindOneParams): Promise<void> {
+  public async deleteTask(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<void> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.deleteTask(task);
   }
 
@@ -98,8 +120,10 @@ export class TasksController {
   async addLabels(
     @Param() { id }: FindOneParams,
     @Body() label: CreateTaskLabelDto[],
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     return await this.tasksService.addLabels(task, label);
   }
 
@@ -108,8 +132,16 @@ export class TasksController {
   async removeLabels(
     @Param() { id }: FindOneParams,
     @Body() labelNames: string[],
+    @CurrentUserId() userId: string,
   ): Promise<void> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.removeLabels(task, labelNames);
+  }
+
+  private checkTaskOwnership(task: Task, userId: string): void {
+    if (task.userId !== userId) {
+      throw new ForbiddenException('You do not have access to this task');
+    }
   }
 }
